+++ app/src/main/java/com/android/boilerplate/roomdatabase/MainActivity.kt
@@ -3,27 +3,55 @@ package com.android.boilerplate.roomdatabase
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
+import androidx.activity.viewModels
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import androidx.room.Room
 import com.android.boilerplate.roomdatabase.ui.theme.RoomDatabaseTheme
 
 class MainActivity : ComponentActivity() {
+
+    private val db by lazy {
+        Room.databaseBuilder(
+            applicationContext,
+            ContactDatabase::class.java,
+            "contacts.db"
+        ).build()
+    }
+
+    private val viewModel by viewModels<ContactViewModel>(
+        factoryProducer = {
+            object : ViewModelProvider.Factory {
+                override fun <T : ViewModel> create(modelClass: Class<T>): T {
+                    return ContactViewModel(db.dao) as T
+                }
+            }
+        }
+
+    )
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
             RoomDatabaseTheme {
-                // A surface container using the 'background' color from the theme
+
+                val state by viewModel.state.collectAsState()
+
+                ContactScreen(state = state, onEvent = viewModel::onEvent)
+
+                /*// A surface container using the 'background' color from the theme
                 Surface(
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
                     Greeting("Android")
-                }
+                }*/
             }
         }
     }